Corregido problema con conversion de entero a double de la canti dad optima de morrales

+++ Morrales/src/morrales/UI/LlenarTablas.java
@@ -200,7 +200,7 @@ ResolverProblema resolver;
            
     }
     
-    public double getNumeroOptimoMorrales(){ 
+    public int getNumeroOptimoMorrales(){ 
        return resolver.getCantOptimaMorrales();
     }
